Worked on implementing the graph-based cleaning problem.py
A lot of work to do, but we are able to create a graph and assign the dirt to the map based on that graph.

Still to do is working on agent logic for solving the problem (tracking the locations of each dirt, which ones have been cleaned and which ones are now available to clean, etc.).  This will require more robust state tracking (TBD).